Handle index + count overflows

+++ src/libraries/System.Collections.Immutable/src/System/Collections/Immutable/ImmutableArray_1.cs
@@ -174,7 +174,7 @@ namespace System.Collections.Immutable
             }
 
             Requires.Range(startIndex >= 0 && startIndex <= self.Length, nameof(startIndex));
-            Requires.Range(count >= 0 && startIndex + count <= self.Length, nameof(count));
+            Requires.Range(count >= 0 && (uint)(startIndex + count) <= (uint)self.Length, nameof(count));
 
             equalityComparer ??= EqualityComparer<T>.Default;
             if (equalityComparer == EqualityComparer<T>.Default)
